% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize}
\alias{normalize}
\alias{is_normalized}
\title{Scale a vector to unit length}
\usage{
normalize(x)

is_normalized(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
\code{normalize()} returns \code{x} divided by its Euclidean norm.
\code{is_normalized()} returns a single \code{TRUE} or \code{FALSE}, testing unit
length up to \code{\link[=all.equal]{all.equal()}} tolerance.
}
\description{
Scale a vector to unit length
}
\examples{
normalize(c(3, 4))
is_normalized(normalize(c(3, 4)))
}
