% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{v2v_rotation_matrix}
\alias{v2v_rotation_matrix}
\title{Rotation matrix taking one unit vector onto another}
\usage{
v2v_rotation_matrix(x, y)
}
\arguments{
\item{x, y}{Unit vectors of the same length. Both are checked with
\code{\link[=is_normalized]{is_normalized()}}.}
}
\value{
A square rotation matrix \code{R} of side \code{length(x)} such that
\code{drop(x \%*\% R)} equals \code{y}.
}
\description{
Vector-to-vector formulation, see
\url{https://en.wikipedia.org/wiki/Rotation_matrix#Vector_to_vector_formulation}.
}
\examples{
R <- v2v_rotation_matrix(c(1, 0, 0), normalize(c(1, 1, 1)))
drop(c(1, 0, 0) \%*\% R)
}
